refactor(models): keep pull progress internal
