+++ pkg/services/models/internal/services/assets/internal/service/prepare_test.go
@@ -1,6 +1,7 @@
 package service
 
 import (
+	"bytes"
 	"context"
 	"crypto/sha256"
 	"encoding/hex"
@@ -681,5 +682,42 @@ func TestPrepareGenericAssetsFetchesManifestWhenRequirementsAreOmitted(t *testin
 	result, err := service.PrepareModelAssets(context.Background(), models.PrepareModelAssetsRequest{Scope: scope, Reference: models.ModelReference{NameOrURI: "hf://owner/repo@" + genericTestRevision}})
 	if err != nil || result.Outcome != models.AssetPreparationPrepared || len(result.Asset.Artifacts) != 1 || result.Asset.Artifacts[0].Name != "weights.bin" || result.Asset.Artifacts[0].SHA256 != sha256Hex(body) {
 		t.Fatalf("manifest-discovered result = %#v", result)
+	}
+}
+
+type progressWriteCloser struct {
+	bytes.Buffer
+	closed bool
+}
+
+func (writer *progressWriteCloser) Close() error {
+	writer.closed = true
+	return nil
+}
+
+func TestCopyStagedAssetReportsTransferredBytesThroughRequestObserver(t *testing.T) {
+	var observations []pullsupport.ProgressObservation
+	ctx := pullsupport.WithProgressObserver(context.Background(), func(observation pullsupport.ProgressObservation) {
+		observations = append(observations, observation)
+	})
+	progress := newAssetProgress(ctx, "voice", "model.bin", int64(len("payload")))
+	var output progressWriteCloser
+
+	written, _, err := copyStagedAssetWithProgress(
+		ctx, &output, strings.NewReader("payload"), "model.bin", progress,
+	)
+	if err != nil {
+		t.Fatalf("copyStagedAssetWithProgress() error = %v", err)
+	}
+	if written != int64(len("payload")) || output.String() != "payload" || !output.closed {
+		t.Fatalf("copy result = written:%d output:%q closed:%t", written, output.String(), output.closed)
+	}
+	if len(observations) < 2 {
+		t.Fatalf("progress observations = %#v, want initial and transferred observations", observations)
+	}
+	last := observations[len(observations)-1]
+	if last.ModelName != "voice" || last.Artifact != "model.bin" ||
+		last.TransferredBytes != int64(len("payload")) || last.TotalBytes != int64(len("payload")) {
+		t.Fatalf("last progress observation = %#v", last)
 	}
 }
